backend/requirements.txt
Django>=6.0
djangorestframework
python-dotenv
psycopg2-binary
